chore(ui): Move menu-data to /lib
Die Datei menu-data.ts enthält die Funktionen menuData und appTextColors für MainNav und MobileNav. Die Datei wurde in den /lib-Ordner verschoben, da es sich hier um Funktionen und nicht um Komponenten handelt.
